requirements.txt
pyarrow
google-cloud-storage
pandas
pytest